(index) Recreated index page
For historical purposes, it may be helpful to know that the code for this
partial was adapted from the code I wrote in the following files:

/application/views/header.php
/application/views/footer.php

Since we don't need CodeIgniter anymore (thanks to Firebase), I can
move these from CodeIgniter into a file with just raw HTML and Javascript.
This will be the base of the MyHonors application.