(CitizenshipCtrl) Fix bug where counters would increment when RSVP was made
This bug was occuring because the $watch was on $rootScope.profile, which was
modified everytime an event was made since the user_profiles Firebase location
also contains RSVP information now. Then, after any RSVP was made, the tally
would increment everything once more. The main problem was that we weren't
clearing the current tally early enough in the process. But some other
simplifications were made also.