(Firebase) Add FirebaseCollection service
This will allow us to easily interact with lists in Firebase. It abstracts
away the difficulties of needing to watch the child_added, child_removed,
and child_changed events. Instead, all we need to do is provide the collection
with the Firebase reference and explicitly call the add(), remove(), and
update() functions on the scope object.

For more info, see the Explicit Synchronization section
at https://github.com/firebase/angularFire