(EventView) Add sorting options for comments
We added a 'date' property to all new comments so they can be sorted
accordingly. It's important to know that adding a negative symbol in front of
any orderBy predicate will reverse the order of it. See the AngularJS docs for
more info on that: http://docs.angularjs.org/api/ng.filter:orderBy

Also, we're including a sort by "Best" option for the future, when we add some
point system to the comments (kudos, upvotes, likes, etc.).