(FirebaseCollection) Add option to grab meta data from a denormalized reference
Since a large part of our schema consists of collections with 'someId: true',
'anotherId: true', etc. it's helpful to automatically link those specific
entries with their actual data. This commit allows us to automatically link
that 'meta' data to the collection, given the fact that we know we're dealing
with a denormalized collection of property names.

Also see: https://www.firebase.com/blog/2013-04-12-denormalizing-is-normal.html