Hide extra features not needed for ARCH release
Since we need to immediately release the ARCH module, but don't want to
expose any of the other features, we simply hide the navigation to other areas
of the site. Yes, users can still access the other routes because we only have
a primative user-access-level system right now, but there isn't any sensitive
information anywhere (we're just using dummy data for testing, and even if we
used real data, users would only be able to see what events we have coming up
in Fall).

In the near future, we will have a better user-access control mechanism that
will allow us to handle these situations very easily and securely.