(ui-bootstrap) Add custom calendar directive
This is totally based off of the ui-bootstrap datepicker directive. It uses
a modified HTML template and adds an events attribute to the directive that
binds to an array in the scope.

I plan to submit this upstream once I generalize it a little more. For
example, the scope.getEvents() function uses a comparison that is exclusive
to our internal data structure.