fix(EventAdd/EventEdit) Fix small bugs
Just a bunch of small bugs.

One interesting one was an obscure error that would only occur if you were
adding an event between 11:00pm and 11:59pm.

The `date.ends.time` field would be auto-populated with the correct time
(1 hour ahead, so 12:XX am) but the `date.ends.date` field wasn't getting
pushed an hour ahead. The result was that the start date was at 11:XX pm
on the current date, and the end date was 12:XX am on the current date.

Since an event can't end before it starts (thanks to our Firebase Security
Rules) it would fail when added.