fix(updateNameData) Use keyname to support profiles without "pid" property
Previously, we were relying on the fact that every profile had a "pid"
property. Not all students have logged in before, however, so in reality not
every profile has that property.

This commit makes this dev operation more robust by using the pid stored in
the keyname (which we are guaranteed to have).