fix(events and citizenship) Limit number of attendants or RSVPs shown in event, show name of events attended in citizenship page
This commit fixes two things. First, it limits the amount of attendants or RSVPs initially shown in an events page. The user can then click 'Show all' to see all attendants. The user can then click 'Show less' to see only a few. The second fix consists of showing the names of the events attended by the user in their citizenship page.

Fixes #27 and resolves #30