Added Bootstrap js, renamed config.js
Added Bootstrap.min.js so that Javascript modules that might be added in the future, that are included in Bootstrap's framework,
can work without a hitch. At this point, it was added because Bootstrap's mobile media query navbar, which auto-magically collapses
into a dropdown menu, wouldn't drop up without this script. Also, config.js was renamed from config.sample.js so that when viewing
an event, data is properly rendered.