Added <div class="row"> back
At first, for some reason, the website wasn't collapsing properly when a breakpoint was hit. I thought that
this was because of the row class, but then found out that it was due to container-fluid's width not being set
to width:100% (with a set max-width in style so that it can go do this to a maximum width of 1150px).