Added CSS rules to alleviate styling problems
Increased the min-width to take into consideration the fact that padding is counted against min-width (making an element's true width = min-width - padding
instead of true width = min-width). This was alleviated by adding the padding on the min-width. Background-size was also changed to contain instead of
cover so that the background image would not go out of bounds beyond the element's dimensions.