Changed max-width, width:100% - container-fluid
Made container-fluid have a max-width of 1170px instead of previous 1100px max-width. This is so that
we can match the container's inner width + 30px padding - container and container-fluid will now, at
their max-width, practically look the same. Different is, one is collapsible, and the other is fixed.
Also, the width of container-fluid was made 100% because of rendering problems, one discovered when
working on the events page - the rows were not collapsing to 100% of the screen size, which was found
as the container-fluid parent wasn't collapsing to 100% of the window's screen size.