Fixed indentation and Attendance display
There was a glitch where the individuals who attended the event were not showing up correctly in a list, which spanned the whole row. Turns out that we still had an old span3 class instead of a col-md-3. Also indented the code for the new markup of modals. Still must fix why the modal directive isn't showing when activated by the delete a href link.