fix(internships) Fixed how internships are ordered and displayed
Everytime an internship is created or edited, the time and date when that occurs is saved. That date is then used as the priority for that internships node in Firebase. That way the latest added and edited internships are shown first in the dashboard as well as the internships pages."