New Points System
New MyHonors Points System based on various Honors Events including
Roomswipes. This pushed for the creation of a stand-alone
CitizenshipService which only gets information about the user's
citizenship once, unless the user refreshes the page or logs-in/out.
Changed coding on DashboardCtrl and CitizenshipCtrl to fit these new
changes, and made some layout changes to citizenship and dashboard html
pages to fit this new citizenship layout.