Citizenship: Variable Objectification and bug fix
In CitizenshipCtrl, instead of declaring multiple scope variable and thus,
multiple watches, I changed to only declaring one object that corresponds
to the citizenship variable. The contents of this object are changed
accordingly as the data is loaded. Also, there was a bug in
CitizenshipService where if a user has more than 10 visits, he/she gets 2
points. Resolved that issue.