Angular scope watch vs scope watchCollection
Firebase takes a while to load. This is causing a problem because
watchCollection only "shallow watches" the object by checking for
additions and deletions of the object. We need to deep watch the object by
also checking for edits on the object itself (changes to properties). This
means to implement scope watch instead of just scope.watchCollection.