Revert to scope.watchCollection and final fix
Finally fixed the problem with CitizenshipPoints not populating as well as
the events list attended by the student - the reason was,
firebase-collections library was returning up an object that is being
watched by firebase with undefined properties. I fixed this by ending the
angular.forEach execution early by returning an empty statement if there
is no name or types property for the event.