MyHonors Events Administration Fixes
When a spot opens up, the earliest registered student in the waiting list
is now the one that is pushed into the RSVP list (not some random
student). Event attendees are now organized by their date registered. We
also fixed the garbled message when the student is added into RSVP list
(fixed update-rsvps-from-waitinglist.php).