fix(events) Fixed RSVP-Waiting List prob
When transfering from the waiting list to the RSVPs, the last people that joined the list were being transferred first, instead of the first people who joined it. Alastair's fixed this with usort() and and did a small fix that made the rest of the code work with usort() (since it modified the array it sorts).