Fix: Non-Honors students can log-in
Non Honors students can log in because an auth_token is being added to the
webStorage, loginSuccess is being set to true, and the login defer is
being resolved even befrore we check if the student is an Honors College
student or not. The check if a student is an Honors College student
if(snapshot.val() === null)) returns the function execution of the
ref.once('value') early and this is the reason behind why such was
implemented. To do the 3-block authentication processes (webStorage.add,
loginSuccess, and defer.resolve) correctly, we must code them after this
early exit fallback.