MyHonors Profiles Fix
Successfully fixed the MyHonors Student Profile problems, which mainly
concerned with file uploads. For some reason, whenever you try uploading a
file, whether it be your profile picture or an image for a project, the
file cannot be viewed. I had to manually CHMOD all the nesting directories
and the file itself that was placed deep inside such directories, doing so
in both file-upload and crop-image PHP files. I also fixed the "The path
to the file is not properly formatted." problem whenever one deletes
files, which failed to delete the deleted project's assets even when given
the correct path locations.