fix,feat(events) Fixed waiting list sort problem. Added isClubMod privilege
This commit fixes an issue with the way students in the waiting list were being sorted in order to be moved to the RSVPs. The sorting was being done in reverse, i.e., the last people who joined the waiting list were being moved to the RSVPs first. Alastair fixed the sorting using usort and I just made a bug fix so that it works with the existing code. This fix is needed because usort modified the actual array and the PID of each student is deleted.

This commit also adds the isClubMod privilige in Firebase and it allows Club mods to swipe people for the events of that club. This feature was quickly added so it is not working completely correct. Right now, a club mod for one club ias able to swipe for another club. This needs to be addressed.